refactor(SE-69 register): Refactor register page
