+++ app/auth/components/ui/CustomAlert.tsx
@@ -1,6 +1,6 @@
-import React, { useEffect } from "react";
 import { Alert, AlertDescription, AlertTitle } from "@/components/ui/alert";
 import { Terminal } from "lucide-react";
+import React from "react";
 
 export interface AlertProps {
   className?: string;
@@ -12,19 +12,10 @@ export interface AlertProps {
 
 const CustomAlert: React.FC<AlertProps> = ({
   className,
-  show,
   title,
   message,
   onClose,
 }) => {
-  // useEffect(() => {
-  //   const timer = setTimeout(() => {
-  //     onClose();
-  //   }, 10000); // Auto hide after 5 seconds
-  //
-  //   return () => clearTimeout(timer);
-  // }, [onClose]);
-
   return (
     <Alert className={className} onClick={onClose}>
       <Terminal className="h-4 w-4" />
